Changes to allow panorama to compile with a newer compiler (gcc-3.0). These changes mostly involved namespaces, but some involved implicit conversion from integral types in calling functions overloaded for both floats and doubles.
Also, I added support for freetype2, while (hopefully) maintaining backwards compatibility for people with the older freetype libraries.  Someone should test the older libraries to see if they still work.

+++ plugins/image_filter/lens_flare/flare.cpp
@@ -102,7 +102,7 @@ void TFlare::setupCorrection (size_t zNOEDGES, float fROTATION)
 
   fROTATION -= (float) (zNOEDGES - 2) * (PI / 2.0f) / (float) zNOEDGES;
 
-  fRotation = fmod (fROTATION, 2.0f * PI);
+  fRotation = fmod (double(fROTATION), double(2.0f * PI));
 
   fArcLength = (2.0f * PI) / (float) zNOEDGES;
